--- a/Mage/src/main/java/mage/cards/repository/CardScanner.java
+++ b/Mage/src/main/java/mage/cards/repository/CardScanner.java
@@ -28,11 +28,13 @@
 
 package mage.cards.repository;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import mage.cards.*;
 import org.apache.log4j.Logger;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 /**
  *
@@ -40,41 +42,51 @@ import org.apache.log4j.Logger;
  */
 public class CardScanner {
 
-    public static boolean scanned = false;
-
     private static final Logger logger = Logger.getLogger(CardScanner.class);
+    public static boolean scanned = false;
 
     public static void scan() {
         if (scanned) {
             return;
         }
         scanned = true;
+        final int availableProcessors = Runtime.getRuntime().availableProcessors();
+        ExecutorService exec = Executors.newFixedThreadPool(availableProcessors);
 
         List<CardInfo> cardsToAdd = new ArrayList<>();
 
         for (ExpansionSet set : Sets.getInstance().values()) {
-            ExpansionRepository.instance.add(new ExpansionInfo(set));
+            exec.submit(new Runnable() {
+                @Override
+                public void run() {
+                    ExpansionRepository.instance.add(new ExpansionInfo(set));
+                }
+            });
         }
-        ExpansionRepository.instance.setContentVersion(ExpansionRepository.instance.getContentVersionConstant());
 
+        ExpansionRepository.instance.setContentVersion(ExpansionRepository.instance.getContentVersionConstant());
         for (ExpansionSet set : Sets.getInstance().values()) {
-            for (ExpansionSet.SetCardInfo setInfo : set.getSetCardInfo()) {
-                if (CardRepository.instance.findCard(set.getCode(), setInfo.getCardNumber()) == null) {
-                    Card card = CardImpl.createCard(setInfo.getCardClass(),
-                            new CardSetInfo(setInfo.getName(), set.getCode(), setInfo.getCardNumber(),
+            exec.submit(new Runnable() {
+                @Override
+                public void run() {
+                    for (ExpansionSet.SetCardInfo setInfo : set.getSetCardInfo()) {
+                        if (CardRepository.instance.findCard(set.getCode(), setInfo.getCardNumber()) == null) {
+                            Card card = CardImpl.createCard(setInfo.getCardClass(),
+                                    new CardSetInfo(setInfo.getName(), set.getCode(), setInfo.getCardNumber(),
                                             setInfo.getRarity(), setInfo.getGraphicInfo()));
-                    if (card != null) {
-                        cardsToAdd.add(new CardInfo(card));
-                        if (card instanceof SplitCard) {
-                            SplitCard splitCard = (SplitCard) card;
-                            cardsToAdd.add(new CardInfo(splitCard.getLeftHalfCard()));
-                            cardsToAdd.add(new CardInfo(splitCard.getRightHalfCard()));
+                            if (card != null) {
+                                cardsToAdd.add(new CardInfo(card));
+                                if (card instanceof SplitCard) {
+                                    SplitCard splitCard = (SplitCard) card;
+                                    cardsToAdd.add(new CardInfo(splitCard.getLeftHalfCard()));
+                                    cardsToAdd.add(new CardInfo(splitCard.getRightHalfCard()));
+                                }
+                            }
                         }
                     }
                 }
-            }
+            });
         }
-
         if (!cardsToAdd.isEmpty()) {
             logger.info("Cards need storing in DB: " + cardsToAdd.size());
             CardRepository.instance.addCards(cardsToAdd);