First loop speed incresed from 35seconds to 19 seconds  on my PC(4 core, 24GB RAM, Windows 8)
Second loop speed increased from 5000ms to 20ms