@sotovdev 2016-12-12
Also i have a question - do you support Java 8 ? Or just Java 7 ?
_that's my very first github commit, so sorry for any inconvinience_